feat(cs): move audit test cases to appropriate place
In addition simplifying bean configuration, making it easier to test against original AuditEventLoggingFacade

Refs: XRDDEV-2243
