+++ src/security-server/admin-service/application/src/test/java/org/niis/xroad/restapi/config/audit/MockableAuditEventLoggingFacade.java
@@ -3,17 +3,17 @@
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -32,33 +32,23 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
-import java.util.Map;
-
 /**
- * We probably need common-rest-api-test dependency for these
+ * We probably need common-rest-api-test dependency for these.
+ * <p>
+ * DEPRECATED. Kept in place until effected tests are fixed.
  */
 @Primary
 @Component
-@Profile({ "test"})
-public class MockableAuditEventLoggingFacade extends AuditEventLoggingFacade {
+@Deprecated
+@Profile({"test"})
+public class MockableAuditEventLoggingFacade extends AuditEventLoggingFacadeImpl {
 
     @Autowired
     public MockableAuditEventLoggingFacade(UsernameHelper usernameHelper,
-            RequestScopedAuditDataHolder requestScopedAuditDataHolder,
-            RequestHelper requestHelper,
-            SecurityHelper securityHelper) {
+                                           RequestScopedAuditDataHolder requestScopedAuditDataHolder,
+                                           RequestHelper requestHelper,
+                                           SecurityHelper securityHelper) {
         super(usernameHelper, requestScopedAuditDataHolder, requestHelper, securityHelper);
     }
 
-    @Override
-    public void callAuditLoggerLogSuccess(RestApiAuditEvent event, String user, Map<String, Object> data,
-            String auth, String url) {
-        super.callAuditLoggerLogSuccess(event, user, data, auth, url);
-    }
-
-    @Override
-    public void callAuditLoggerLogFailure(RestApiAuditEvent event, String user, String reason,
-            Map<String, Object> data, String auth, String url) {
-        super.callAuditLoggerLogFailure(event, user, reason, data, auth, url);
-    }
 }
